♻️ Refactor: 코드리뷰에 따라 순환참조 해소 등 문제 해결
