+++ src/shared/components/header/header.types.ts
@@ -9,8 +9,6 @@ export interface HeaderProfileAvatarProps {
 
 export interface HeaderProps {
   user?: HeaderUser;
-  hasNotification?: boolean;
-  onNotificationClick?: () => void;
   onProfileClick?: () => void;
   isProfileMenuOpen?: boolean;
   profileMenuId?: string;
